build(website): add an @website annotation parser to core stories
This change does a couple of things all related to generated data for
the website to pull snippits of html from core stories into the code
tab.

First, it adds a parser script that uses ts-morph (an ast abstraction
package for the typescript ast. The script can find functions that are
annotated with /** @website **/ and identify only the html template for
that story.

Second, this change modifies stories to export named functions instead
of const lambda functions.

Third, output is saved to the websites data folder to be consumed there.

Signed-off-by: Matt Hippely <mhippely@vmware.com>